Fix user input bugs, add conditionals.

+++ bazel-up.sh
@@ -1,24 +1,39 @@
-sudo apt update && sudo apt upgrade
+sudo apt update && sudo apt upgrade -y
 
-sudo apt install curl gnupg
+sudo apt install -y curl gnupg
 curl -fsSL https://bazel.build/bazel-release.pub.gpg | gpg --dearmor > bazel.gpg
 sudo mv bazel.gpg /etc/apt/trusted.gpg.d/
 echo "deb [arch=amd64] https://storage.googleapis.com/bazel-apt stable jdk1.8" | sudo tee /etc/apt/sources.list.d/bazel.list
 
 
-sudo apt update && sudo apt install bazel
-sudo apt update && sudo apt full-upgrade
-sudo apt install bazel-1.0.0
+sudo apt update && sudo apt install -y bazel
+sudo apt update && sudo apt full-upgrade -y
+sudo apt install -y bazel-1.0.0
 
-sudo ln -s /usr/bin/bazel-1.0.0 /usr/bin/bazel
+if [ ! -f /usr/bin/bazel-1.0.0 ]
+then
+  echo "File /usr/bin/bazel-1.0.0 does not exist."
+  sudo ln -s /usr/bin/bazel-1.0.0 /usr/bin/bazel
+else
+  echo "File /usr/bin/bazel-1.0.0 exists."
+fi
+  
 bazel --version  # 1.0.0
 
 
-sudo apt install g++ unzip zip
+sudo apt install -y g++ unzip zip
 
-chmod +x bazel-<version>-installer-linux-x86_64.sh
-./bazel-<version>-installer-linux-x86_64.sh --user
+if [ ! -f bazel-6.0.0-pre.20211025.1-installer-linux-x86_64.sh ]
+then
+  echo "No Bazel installer, downloading:"
+  wget https://github.com/bazelbuild/bazel/releases/download/6.0.0-pre.20211025.1/bazel-6.0.0-pre.20211025.1-installer-linux-x86_64.sh
+else
+  echo "Existing Bazel installer found."
+fi
+
+sudo chmod +x bazel-6.0.0-pre.20211025.1-installer-linux-x86_64.sh
+
+bash bazel-6.0.0-pre.20211025.1-installer-linux-x86_64.sh --user
 
 echo 'export PATH="$PATH:$HOME/bin"' >> ~/.bashrc
-
